+++ CHANGELOG.md
@@ -9,6 +9,7 @@ New features:
 
 * Image processing:
   * Added Total variation denoising algorithm (TV Chambolle)
+  * Added Bilateral filter denoising algorithm
 * Added support for XYZ image files
 
 Bug fixes: